requirements.txt
django
djangorestframework
psycopg2
drf-yasg
djangorestframework-simplejwt
djangorestframework-jwt 
python-jose
ConfigParser
asgiref
cffi
click
cryptography
h11
pycparser
pydantic
starlette
uvicorn
fastapi
requests
psycopg2
Pillow
reportlab
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
